Example for named WebP transform. And Spelling.
For creating a WebP <source> adding the format parameter is not sufficient. A dedicated named transform is necessary. The documentation correctly showed a different transform name but did not explicitly tell to create such an transform. For easier understanding the info is now added and an example is shown.

A couple of spelling errors are corrected.